Fixed some memory problems in the Stokes solver.
Some arrays were not allcated with enough elements (off by 1). Writing to the
end of those arrays can potentially overwrite other parts of the memory. In
practice, it seems that this bug didn't affect the solver at all, but this is
a subtle bug to detect.

This bug was in CitcomS pre-1.0 version and is likely in other versions of
citcom family as well.